+++ Items/Weapons/Melee/Helios.cs
@@ -1,14 +1,7 @@
 ﻿
 using Microsoft.Xna.Framework;
 using Stellamod.Helpers;
-using Stellamod.Items.Harvesting;
-using Stellamod.Items.Materials;
-using Stellamod.Items.Ores;
-using Stellamod.Particles;
-using Stellamod.Projectiles;
 using Stellamod.Projectiles.Magic;
-using Stellamod.Projectiles.Slashers;
-using Stellamod.Projectiles.Slashers.Gutinier;
 using Stellamod.Projectiles.Slashers.Helios;
 using System.Collections.Generic;
 using Terraria;
@@ -87,11 +80,11 @@ namespace Stellamod.Items.Weapons.Melee
             int dir = AttackCounter;
             if (player.direction == 1)
             {
-                player.GetModPlayer<CorrectSwing>().SwingChange = (int)AttackCounter;
+                player.GetModPlayer<CorrectSwing>().SwingChange = AttackCounter;
             }
             else
             {
-                player.GetModPlayer<CorrectSwing>().SwingChange = (int)AttackCounter * -1;
+                player.GetModPlayer<CorrectSwing>().SwingChange = AttackCounter * -1;
 
             }
             AttackCounter = -AttackCounter;
