+++ NPCs/Bosses/Verlia/Projectiles/FrostShot.cs
@@ -1,16 +1,11 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using ParticleLibrary;
-using Stellamod.Dusts;
-using Stellamod.Particles;
-using System.Collections.Generic;
 using Terraria;
-using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace Stellamod.NPCs.Bosses.Verlia.Projectiles
 {
-	public class FrostShot : ModProjectile
+    public class FrostShot : ModProjectile
 	{
 		public override void SetStaticDefaults()
 		{
@@ -56,7 +51,7 @@ namespace Stellamod.NPCs.Bosses.Verlia.Projectiles
 				float speedYabc = -Projectile.velocity.Y * Main.rand.Next(0, 0) * 0.00f + Main.rand.Next(0, 0) * 0.0f;
 
 
-				Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.position.X + speedXabc - 20, Projectile.position.Y + speedYabc - 20, speedXabc * 0, speedYabc * 0, ModContent.ProjectileType<FrostShotIN>(), (int)(Projectile.damage * 0), 0f, Projectile.owner, 0f, 0f);
+				Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.position.X + speedXabc - 20, Projectile.position.Y + speedYabc - 20, speedXabc * 0, speedYabc * 0, ModContent.ProjectileType<FrostShotIN>(), Projectile.damage * 0, 0f, Projectile.owner, 0f, 0f);
 				Timer = 0;
