+++ Projectiles/SalfaCircle.cs
@@ -4,7 +4,7 @@ using Terraria.ModLoader;
 
 namespace Stellamod.Projectiles
 {
-	public class SalfaCircle : ModProjectile
+    public class SalfaCircle : ModProjectile
 	{
 		public override void SetStaticDefaults()
 		{
@@ -31,7 +31,7 @@ namespace Stellamod.Projectiles
 		}
 		public override Color? GetAlpha(Color lightColor)
 		{
-			return new Color(26, 38, 22, 0) * (1f - (float)Projectile.alpha / 50f);
+			return new Color(26, 38, 22, 0) * (1f - Projectile.alpha / 50f);
 		}
 		public override void AI()
 		{
